pyproject.toml
[tool.poetry]
name = "product-migration"
version = "0.2.0"
description = "Tool to migrate from OpenProject to Plane"
authors = ["TheDarkNine <thedarknine@proton.me>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v -ra -q"
#log_cli = true
minversion = "6.0"
# Specify a custom directory for pytest cache
cache_dir = "./.tmp/.pytest_cache"
# Env variables
env_override_existing_values = 1
env_files = [
    '.env.test',
]

[tool.numpydoc_validation]
checks = [
    "all",   # report on all checks, except the below
    "EX01",
    "SA01",
    "ES01",
]
convention = "google"
exclude = []

[tool.ruff]
line-length = 100
exclude = ["__init__.py"]
cache-dir = ".tmp/ruff_cache"
[tool.ruff.lint]
select = ["E", "F", "D"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.28.0"
python-dotenv = "^1.0.1"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
pytest-httpx = "^0.35.0"
pytest-dotenv = "^0.5.2"
pylint = "^3.3.2"
black = "^24.10.0"
arrow = "^1.3.0"
rich = "^13.9.4"
pydantic = "^2.10.3"
toml = "^0.10.2"
numpydoc = "^1.8.0"
ruff = "^0.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
